pyproject.toml
[project]
name = "sistemundigitalkelabkkqsmkdd1"
version = "0.1.0"
description = "Sistem Undigital Kelab KKQ SMKDD1 port from PHP to Python using Flask"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "flask>=3.1.3",
    "flask-login>=0.6.3",
    "flask-migrate>=4.1.0",
    "flask-sqlalchemy>=3.1.1",
    "python-dotenv>=1.2.2",
]
